+++ src/adapter/next/handler.ts
@@ -1,10 +1,9 @@
-import {
-  AvailableDrivers,
-  BackendOptions,
-  createBackend
-} from 'alinea/backend/api/CreateBackend'
+import {BackendOptions, createBackend} from 'alinea/backend/api/CreateBackend'
 import {Backend} from 'alinea/backend/Backend'
-import {createHandler as createCoreHandler} from 'alinea/backend/Handler'
+import {
+  createHandler as createCoreHandler,
+  HandlerHooks
+} from 'alinea/backend/Handler'
 import {JWTPreviews} from 'alinea/backend/util/JWTPreviews'
 import {cloudBackend} from 'alinea/cloud/CloudBackend'
 import {Entry} from 'alinea/core/Entry'
@@ -15,16 +14,33 @@ import {devUrl, requestContext} from './context.js'
 type Handler = (request: Request) => Promise<Response>
 const handlers = new WeakMap<NextCMS, Handler>()
 
-export function createHandler<Driver extends AvailableDrivers>(
+export interface NextHandlerOptions extends HandlerHooks {
+  cms: NextCMS
+  backend?: BackendOptions | Backend
+}
+
+export function createHandler(cms: NextCMS): Handler
+export function createHandler(options: NextHandlerOptions): Handler
+/** @deprecated */
+export function createHandler(
   cms: NextCMS,
-  backend: BackendOptions<Driver> | Backend = cloudBackend(cms.config)
-) {
-  if (handlers.has(cms)) return handlers.get(cms)!
-  const api = 'database' in backend ? createBackend(backend) : backend
-  const handleBackend = createCoreHandler(cms, api)
+  backend: BackendOptions | Backend
+): Handler
+export function createHandler(
+  input: NextCMS | NextHandlerOptions,
+  backend?: BackendOptions | Backend
+): Handler {
+  const options = input instanceof NextCMS ? {cms: input, backend} : input
+  const providedBackend = options.backend ?? cloudBackend(options.cms)
+  if (handlers.has(options.cms)) return handlers.get(options.cms)!
+  const api =
+    'database' in providedBackend
+      ? createBackend(providedBackend)
+      : providedBackend
+  const handleBackend = createCoreHandler({...options, backend: api})
   const handle: Handler = async request => {
     try {
-      const context = await requestContext(cms.config)
+      const context = await requestContext(options.cms.config)
       const previews = new JWTPreviews(context.apiKey)
       const {searchParams} = new URL(request.url)
       const previewToken = searchParams.get('preview')
@@ -36,7 +52,7 @@ export function createHandler<Driver extends AvailableDrivers>(
         const info = await previews.verify(previewToken)
         const cookie = await cookies()
         const connection = devUrl()
-          ? await cms.connect()
+          ? await options.cms.connect()
           : handleBackend.connect(context)
         const payload = getPreviewPayloadFromCookies(cookie.getAll())
         const url = await connection.resolve({
@@ -65,6 +81,6 @@ export function createHandler<Driver extends AvailableDrivers>(
       return new Response('Internal server error', {status: 500})
     }
   }
-  handlers.set(cms, handle)
+  handlers.set(options.cms, handle)
   return handle
 }