+++ src/core/Graph.ts
@@ -1,19 +1,17 @@
 import {Root, Workspace} from 'alinea/types'
-import * as cito from 'cito'
 import {Config} from './Config.js'
 import {EntryFields} from './EntryFields.js'
 import {Expr} from './Expr.js'
 import {Condition, Filter} from './Filter.js'
 import {Infer, StoredRow} from './Infer.js'
+import {HasField} from './Internal.js'
 import {Page} from './Page.js'
 import {PreviewRequest} from './Preview.js'
 import {Resolver} from './Resolver.js'
 import {Type} from './Type.js'
-import {hasExact} from './util/Checks.js'
 import {Expand} from './util/Types.js'
 
 export type Location = Root | Workspace | Page | Array<string>
-type EmptyObject = Record<PropertyKey, never>
 
 type FieldsOf<Types> = StoredRow<
   Types extends Type<infer V>
@@ -23,35 +21,74 @@ type FieldsOf<Types> = StoredRow<
     : unknown
 >
 
-export interface RelatedQuery<Selection = unknown, Types = unknown>
-  extends GraphQuery<Selection, Types> {
-  translations?: EmptyObject | {includeSelf: boolean | undefined}
-  children?: EmptyObject | {depth: number | undefined}
-  parents?: EmptyObject | {depth: number | undefined}
-  siblings?: EmptyObject | {includeSelf: boolean | undefined}
-
-  parent?: EmptyObject
-  next?: EmptyObject
-  previous?: EmptyObject
+export interface EdgeTranslations {
+  edge: 'translations'
+  includeSelf?: boolean
 }
 
-type IsRelated =
-  | {translations: EmptyObject | {includeSelf: boolean | undefined}}
-  | {children: EmptyObject | {depth: number | undefined}}
-  | {parents: EmptyObject | {depth: number | undefined}}
-  | {siblings: EmptyObject | {includeSelf: boolean | undefined}}
-  | {parent: EmptyObject}
-  | {next: EmptyObject}
-  | {previous: EmptyObject}
-
-interface ToSelect {
-  [key: string]:
-    | Expr<any>
-    | RelatedQuery<ToSelect | Expr<any>, Type | Array<Type>>
-    | ToSelect
+export interface EdgeChildren {
+  edge: 'children'
+  depth?: number
 }
 
-export type Projection = ToSelect | Expr<any>
+export interface EdgeParents {
+  edge: 'parents'
+  depth?: number
+}
+
+export interface EdgeSiblings {
+  edge: 'siblings'
+  includeSelf?: boolean
+}
+
+export interface EdgeParent {
+  edge: 'parent'
+}
+
+export interface EdgeNext {
+  edge: 'next'
+}
+
+export interface EdgePrevious {
+  edge: 'previous'
+}
+
+export interface EdgeEntries {
+  edge: 'entryMultiple'
+  field: HasField & Expr
+}
+
+export interface EdgeEntry {
+  edge: 'entrySingle'
+  field: HasField & Expr
+}
+
+export type Edge =
+  | EdgeTranslations
+  | EdgeChildren
+  | EdgeParents
+  | EdgeSiblings
+  | EdgeParent
+  | EdgeNext
+  | EdgePrevious
+  | EdgeEntries
+  | EdgeEntry
+
+export type EdgeQuery<
+  Selection = unknown,
+  Types = unknown,
+  Include = unknown
+> = GraphQuery<Selection, Types, Include> & Edge
+
+type SelectContent =
+  | Expr<any>
+  | EdgeQuery<SelectObject | Expr<any>, Type | Array<Type>>
+
+interface SelectObject {
+  [key: string]: Projection
+}
+
+export type Projection = SelectContent | SelectObject
 export type InferProjection<Selection> = InferSelection<Selection>
 
 export interface Order {
@@ -60,28 +97,24 @@ export interface Order {
   caseSensitive?: boolean
 }
 
-type InferSelection<Selection> = Selection extends Expr<infer V>
+type InferSelection<Selection> = Selection extends GraphQuery & Edge
+  ? Expand<AnyQueryResult<Selection>>
+  : Selection extends Expr<infer V>
   ? V
   : {
       [K in keyof Selection]: Selection[K] extends Type<infer V>
         ? Type.Infer<V>
-        : Selection[K] extends Expr<infer V>
-        ? V
-        : Selection[K] extends GraphQuery & IsRelated
-        ? Expand<AnyQueryResult<Selection[K]>>
         : InferSelection<Selection[K]>
     }
 
-type InferResult<Selection, Types, Include> = Selection extends Expr<
-  infer Value
->
-  ? Value
-  : Selection extends undefined
+type InferResult<Selection, Types, Include> = Selection extends undefined
   ? Types extends undefined
     ? EntryFields & (Include extends undefined ? {} : InferSelection<Include>)
     : EntryFields &
         Infer<Types> &
         (Include extends undefined ? {} : InferSelection<Include>)
+  : Selection extends Expr<infer Value>
+  ? Value
   : InferSelection<Selection>
 
 type QueryResult<Selection, Types, Include> = Expand<
@@ -94,9 +127,10 @@ interface CountQuery<Selection, Types, Include>
 }
 type FirstQuery<Selection, Types, Include> =
   | (GraphQuery<Selection, Types, Include> & {first: true})
-  | (GraphQuery<Selection, Types, Include> & {next: EmptyObject})
-  | (GraphQuery<Selection, Types, Include> & {previous: EmptyObject})
-  | (GraphQuery<Selection, Types, Include> & {parent: EmptyObject})
+  | (GraphQuery<Selection, Types, Include> & {edge: 'parent'})
+  | (GraphQuery<Selection, Types, Include> & {edge: 'next'})
+  | (GraphQuery<Selection, Types, Include> & {edge: 'previous'})
+
 interface GetQuery<Selection, Types, Include>
   extends GraphQuery<Selection, Types, Include> {
   get: true
@@ -108,10 +142,10 @@ export type AnyQueryResult<Query extends GraphQuery> = Query extends CountQuery<
   any
 >
   ? number
-  : Query extends FirstQuery<infer S, infer T, infer I>
-  ? QueryResult<S, T, I> | null
   : Query extends GetQuery<infer S, infer T, infer I>
   ? QueryResult<S, T, I>
+  : Query extends FirstQuery<infer S, infer T, infer I>
+  ? QueryResult<S, T, I> | null
   : Query extends GraphQuery<infer S, infer T, infer I>
   ? Array<QueryResult<S, T, I>>
   : unknown
@@ -212,7 +246,7 @@ export interface GraphQuery<
 
 export type SelectionGuard = Projection | undefined
 export type TypeGuard = Type | Array<Type> | undefined
-export type IncludeGuard = ToSelect | undefined
+export type IncludeGuard = SelectObject | undefined
 
 export class Graph {
   #resolver: Resolver
@@ -262,32 +296,8 @@ export class Graph {
   }
 }
 
-const plainObject = cito.type((value): value is object => {
-  return value?.constructor === Object
-})
-const emptySource = plainObject.and(hasExact([]))
-const depthSource = cito
-  .object({depth: cito.number.optional})
-  .and(hasExact(['depth']))
-const includeSource = cito
-  .object({includeSelf: cito.boolean.optional})
-  .and(hasExact(['includeSelf']))
-const translationsSource = cito.object({translations: includeSource})
-const childrenSource = cito.object({children: depthSource})
-const parentsSource = cito.object({parents: depthSource})
-const siblingsSource = cito.object({siblings: includeSource})
-
-const parentSource = cito.object({parent: emptySource})
-const nextSource = cito.object({next: emptySource})
-const prevSource = cito.object({previous: emptySource})
-
 export function querySource(query: unknown) {
   if (query?.constructor !== Object) return
-  if (translationsSource.check(query)) return 'translations'
-  if (childrenSource.check(query)) return 'children'
-  if (parentsSource.check(query)) return 'parents'
-  if (siblingsSource.check(query)) return 'siblings'
-  if (parentSource.check(query)) return 'parent'
-  if (nextSource.check(query)) return 'next'
-  if (prevSource.check(query)) return 'previous'
+  const related = (<Edge>query).edge
+  return related
 }